GUI Components added.
This commit i added a lot of changes to the gui. There needs to be event
handlers for a lot of the components though. layouts have been changed
from absolute to gridlayout/gridbaglayout to be able to resize with the
window. Added a slider to change the value of spacing and a label to
display the current value of spacing. Created separate buttons for
converting the file, saving the file, and previewing the file (event
handlers need to still be configured for the new components). In
addition i separated the components to have text stuff on one side and
pdf stuff on the other.