fixed pdf displaying random spaces in windows, usermanual will now get displayed if the user selects it from the menu bar.
seems that the fonts were the thing that screwed up the pdf displaying.
I'm not really sure the reason behind it but it works fine if u dont use
the icepdf library to display with the previous version. i changed the
encoding from WINANSI to Cp1252 to fix this issue. Needs to be tested on
other OS's to make sure it displays ok