Updated GUI
Updated the look of the gui. the pdf should now be displayed properly...
please test if it works on both platforms. added a temporary pdf to the
about section, the user manual in the manual section in the menu.
TODO:: When provided a new text file reset the gui to default values,
alert boxes, disable buttons if they cannot be used.