Gui has been changed to display the pdf in the page
Removed the display button. Convert button now also dispalys the pdf in
the current frame if it is completed. moved the localDest to the
drawoutput as it makes more sense to be there. modified the method
createPdf and initPdf to require a boolean as a parameter to determine
whether or not to use the localDest.